refactor(embeddings): extract shared model-cache helpers
Move is_model_cached, detect_country, download_via_modelscope, and
bridge_modelscope_to_hf_cache out of pyseekdb_default.py into a new
_model_cache.py module so both pyseekdb_default and huggingface can
share the same CN-mirror download / cache-bridge logic without
duplication. Tests are updated to patch _model_cache instead of
pyseekdb_default internals.
